feat(LT-4621): add device configuration repository

+++ src/Lykke.Snow.Notifications.SqlRepositories/Repositories/DeviceConfigurationRepository.cs
@@ -5,7 +5,6 @@ using Lykke.Snow.Notifications.Domain.Repositories;
 using Lykke.Snow.Notifications.SqlRepositories.Entities;
 using Lykke.Snow.Notifications.SqlRepositories.Exceptions;
 using Microsoft.EntityFrameworkCore;
-using DeviceConfiguration = Lykke.Snow.Notifications.Domain.Repositories.DeviceConfigurationStub;
 
 namespace Lykke.Snow.Notifications.SqlRepositories.Repositories
 {
@@ -22,65 +21,48 @@ namespace Lykke.Snow.Notifications.SqlRepositories.Repositories
             _mapper = mapper;
         }
 
-        public async Task<DeviceConfiguration?> GetByIdAsync(int id)
+        public async Task<DeviceConfiguration?> GetAsync(string deviceId)
         {
             await using var context = _contextFactory.CreateDataContext();
 
             var entity = await context.DeviceConfigurations
                 .Include(x => x.Notifications)
-                .FirstOrDefaultAsync(x => x.Oid == id);
+                .FirstOrDefaultAsync(x => x.DeviceId == deviceId);
             
             if (entity == null)
-                throw new EntityNotFoundException(id);
-            
+                throw new EntityNotFoundException(deviceId);
+
             return _mapper.Map<DeviceConfiguration>(entity);
         }
 
-        public async Task AddAsync(DeviceConfiguration deviceConfiguration)
+        public async Task AddOrUpdateAsync(DeviceConfiguration deviceConfiguration)
         {
             await using var context = _contextFactory.CreateDataContext();
-            
-            var entity = _mapper.Map<DeviceConfigurationEntity>(deviceConfiguration);
-            await context.DeviceConfigurations.AddAsync(entity);
-            
-            try
+
+            var existingEntity = await context.DeviceConfigurations
+                .Include(x => x.Notifications)
+                .FirstOrDefaultAsync(x => x.DeviceId == deviceConfiguration.DeviceId);
+
+            if (existingEntity == null)
             {
-                await context.SaveChangesAsync();
+                await TryAddAsync(context, deviceConfiguration);
+                return;
             }
-            catch(DbUpdateException e)
-            {
-                if(e.ValueAlreadyExistsException())
-                {
-                    throw new EntityAlreadyExistsException(entity: entity);
-                }
-                
-                throw;
-            };
+
+            await TryUpdateAsync(context, deviceConfiguration, existingEntity);
         }
 
-        public async Task UpdateAsync(DeviceConfiguration deviceConfiguration)
+        public async Task RemoveAsync(string deviceId)
         {
             await using var context = _contextFactory.CreateDataContext();
             
-            var entity = _mapper.Map<DeviceConfigurationEntity>(deviceConfiguration);
-            context.DeviceConfigurations.Update(entity);
+            var entity = await context.DeviceConfigurations
+                .Include(x => x.Notifications)
+                .FirstOrDefaultAsync(x => x.DeviceId == deviceId);
 
-            try
-            {
-                await context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException e) when (e.IsMissingDataException())
-            {
-                throw new EntityNotFoundException(entity.Oid);
-            }
-        }
+            if (entity == null)
+                throw new EntityNotFoundException(deviceId);
 
-        public async Task RemoveAsync(int id)
-        {
-            await using var context = _contextFactory.CreateDataContext();
-            
-            var entity = new DeviceConfigurationEntity {Oid = id};
-            context.DeviceConfigurations.Attach(entity);
             context.DeviceConfigurations.Remove(entity);
             
             try 
@@ -89,7 +71,42 @@ namespace Lykke.Snow.Notifications.SqlRepositories.Repositories
             }
             catch(DbUpdateConcurrencyException e) when (e.IsMissingDataException())
             {
-                throw new EntityNotFoundException(id);
+                throw new EntityNotFoundException(deviceId);
+            }
+        }
+
+        private async Task TryAddAsync(NotificationsDbContext context, DeviceConfiguration deviceConfiguration)
+        {
+            var entity = _mapper.Map<DeviceConfigurationEntity>(deviceConfiguration);
+            await context.DeviceConfigurations.AddAsync(entity);
+            try
+            {
+                await context.SaveChangesAsync();
+            }
+            catch (DbUpdateException e)
+            {
+                if (e.ValueAlreadyExistsException())
+                {
+                    throw new EntityAlreadyExistsException(entity: entity);
+                }
+
+                throw;
+            }
+        }
+
+        private async Task TryUpdateAsync(NotificationsDbContext context,
+            DeviceConfiguration deviceConfiguration,
+            DeviceConfigurationEntity existingEntity)
+        {
+            _mapper.Map(deviceConfiguration, existingEntity);
+            context.DeviceConfigurations.Update(existingEntity);
+            try
+            {
+                await context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException e) when (e.IsMissingDataException())
+            {
+                throw new EntityNotFoundException(existingEntity.Oid);
             }
         }
     }
